Progress 25/11/2025
Added a pattern-based cell classifier to sort cells into groups (house, objective, environment) based on previously written color sorting function. Added a rudimentary recursive backtracking algorithm that is capable of finding a functioning path from a house to it's respective objective.

This is the first end-to-end version of this project where the computer is capable of scanning the screen and placing calculated paths onto the game board. Further optimization is needed for both the autonomous environment detection and pathfinding algorithm.

Once optimization is fully complete/functional enough, the next step will be to run the bot to repeat this process periodically so it can autonomously play the game as it progresses. Currently, it only scans the current screen and plays once.