Progress 26/11/2025
Added computer vision detection for the actual connection points of the objectives through brightness detection of cells around objectives. Modularized ROI calculations. Randomized the recursive backtracking (DFS) and made it run 100 times per house to find a more optimal path.

This is the first semi-functioning prototype of the bot. The bot can now autonomously play the beginning levels of the game with moderate success.